Criação de temas

+++ src/main/java/com/generation/blogpessoal1/controller/PostagemController.java
@@ -1,6 +1,7 @@
-package com.generation.blogpessoal1.controller;
+package com.generation.blogpessoal1.controller; // comunicação
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,10 +14,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.generation.blogpessoal1.model.Postagem;
 import com.generation.blogpessoal1.repository.PostagemRepository;
+import com.generation.blogpessoal1.repository.TemaRepository;
 
 import jakarta.validation.Valid;
 
@@ -24,47 +28,61 @@ import jakarta.validation.Valid;
 @RequestMapping("/postagens")
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 public class PostagemController {
-
 	@Autowired
 	private PostagemRepository postagemRepository;
 	
+	@Autowired
+	private TemaRepository temaRepository;
+	
 	@GetMapping
 	public ResponseEntity<List<Postagem>> getAll(){
+		// SELECT * FROM tb_postagens
 		return ResponseEntity.ok(postagemRepository.findAll());
-		//SELECT*FROM tb_postagens;
 	}
-		
-	@GetMapping("/{id}")
+	
+	@GetMapping("{id}")
 	public ResponseEntity<Postagem> getById(@PathVariable Long id){
+		// SELECT * FROM tb_postagens WHERE id = ?
 		return postagemRepository.findById(id)
-			.map(resposta -> ResponseEntity.ok(resposta))
-			.orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
-		
-		//SELECT * FROM tb_postages WHERE i = ?;
+				.map(resposta -> ResponseEntity.ok(resposta))
+				.orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
 	}
+	
 	@GetMapping("/titulo/{titulo}")
 	public ResponseEntity<List<Postagem>> getByTitulo(@PathVariable String titulo){
+		// SELECT * FROM tb_postagens WHERE titulo LIKE "%???%"
 		return ResponseEntity.ok(postagemRepository.findAllByTituloContainingIgnoreCase(titulo));
-		
-		// SELECT * FROM tb_postages WHERE titulo LIKE "
-}
+	}
+	
 	@PostMapping
 	public ResponseEntity<Postagem> post(@Valid @RequestBody Postagem postagem){
-		return ResponseEntity.status(HttpStatus.CREATED)
-				.body(postagemRepository.save(postagem));
+		// INSERT INTO tb_postagens(titulo, texto) VALUES ("???", "???")
+		
+		return temaRepository.findById(postagem.getTema().getId())
+				.map(resposta -> ResponseEntity.status(HttpStatus.CREATED).body(postagemRepository.save(postagem)))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
+	
 	@PutMapping
 	public ResponseEntity<Postagem> put(@Valid @RequestBody Postagem postagem){
-		return ResponseEntity.status(HttpStatus.CREATED)
-				.body(postagemRepository.save(postagem));
-		/*UPDATE tb_postagens*/
-}
-	@DeleteMapping("/{id}")
+		// UPDATE tb_postagens SET titulo = "???", texto = "???", data = "???" WHERE id = ?
+		return postagemRepository.findById(postagem.getId())
+				.map(resposta -> ResponseEntity.ok().body(postagemRepository.save(postagem)))
+				.orElse(ResponseEntity.notFound().build());
+		}
+	
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	@DeleteMapping("{id}")
 	public void delete(@PathVariable Long id) {
-		postagemRepository.deleteById(id);
+		// DELETE FROM tb_postagens WHERE id = ?
 		
-		/*DELETE FROM tb_postagens where id= id*/
+		Optional<Postagem> optional = postagemRepository.findById(id);
+		
+		if (optional.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+		}
+		else {
+			postagemRepository.deleteById(id);
+		}
 	}
 }
-	
-
